+++ src/resample.c
@@ -22,27 +22,27 @@ static boolean reorder_leave_back = FALSE;
 
 LIVES_GLOBAL_INLINE ticks_t q_gint64(ticks_t in, double fps) {
   // quantise timecode to fps
-  if (in > (ticks_t)0) return ((ticks_t)((long double)in / (long double)TICKS_PER_SECOND_DBL * (long double)fps +
-                                           (long double).5) /
-                                 (long double)fps) * (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
-  if (in < (ticks_t)0) return ((ticks_t)((long double)in / (long double)TICKS_PER_SECOND_DBL * (long double)fps -
-                                           (long double).5) /
-                                 (long double)fps) * (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
+  if (in > (ticks_t)0) return ((ticks_t)((double)in / (double)TICKS_PER_SECOND_DBL * (double)fps +
+                                           (double).5) /
+                                 (double)fps) * (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
+  if (in < (ticks_t)0) return ((ticks_t)((double)in / (double)TICKS_PER_SECOND_DBL * (double)fps -
+                                           (double).5) /
+                                 (double)fps) * (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
   return (ticks_t)0;
 }
 
 LIVES_GLOBAL_INLINE ticks_t q_gint64_floor(ticks_t in, double fps) {
-  if (in != (ticks_t)0) return ((ticks_t)((long double)in / (long double)TICKS_PER_SECOND_DBL * (long double)fps) /
-                                  (long double)fps) *
+  if (in != (ticks_t)0) return ((ticks_t)((double)in / (double)TICKS_PER_SECOND_DBL * (double)fps) /
+                                  (double)fps) *
                                  (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
   return (ticks_t)0;
 }
 
 LIVES_GLOBAL_INLINE ticks_t q_dbl(double in, double fps) {
   // quantise (double)in to fps
-  if (in > 0.) return ((ticks_t)((long double)in * (long double)fps + (long double).5) / (long double)fps) *
+  if (in > 0.) return ((ticks_t)((double)in * (double)fps + (double).5) / (double)fps) *
                         (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
-  if (in < 0.) return ((ticks_t)((long double)in * (long double)fps - (long double).5) / (long double)fps) *
+  if (in < 0.) return ((ticks_t)((double)in * (double)fps - (double).5) / (double)fps) *
                         (ticks_t)TICKS_PER_SECOND; // quantise to frame timing
   return (ticks_t)0;
 }
